Updated JSerial to allow flow control to be disabled
Updated JSerial to allow flow control to be disabled. Add the
SerialFlowControlEnabled:

A serial port may use signals in the interface
to pause and resume the transmission of data. For example, a
slow printer might need to handshake with the serial port to
indicate that data should be paused while the mechanism advances a
line. These signals can implemented by hardware or by software.
Use 1 to enabled flow control. Use 0 to disbale flow control.
If flow control is enabled, use SerialSoftwareFlowControl to
configure hardware or software. Flow control is enabled by default.

Fixed #27.