Fixed interpretation attributes truncated to 255 chars
Fixed interpretation attributes being truncated to 255 characters.
Needed to update the ```GUI/src/com/u2d/generated/``` classes to switch
the interpretation from a ```StringEO``` to a ```TextEO```. Also needed
to update the ```GUI/src/org/jts/gui/JAXBtoJmatter/``` classes to
stop truncating the string if it is more than 255 characters.

Finally, needed to update ```GUI/src/org/jts/gui/util/Conversion.java```
because the ```setNonEmptyDescription``` method uses reflection to
get the interpretation. It was casting it to a ```StringEO```. I've
update the method to handle ```StringEO```, ```TextEO```, and ```String```.

Fixed #12.