Fixed JTS failing to import JSIDL with messages that don't have is_command.
Fixed JTS failing to import JSIDL with messages that don't have is_command.
This attribute is optional. No default is specified, but since it was added
in JSIDL 1.1, I'm going to assume the default is false because messages
defined prior to JSIDL were NOT commands.

The JMatter GUI elements don't support optional fields. It would be nice
to have that option in there. For now, the is_command will be added to
all messages.

SIDENOTE:
The JSIDL 1.1 should be updated to indicate a default value for optional
fields.