Fixed #54. Updated code gen to support service id and version.
Fixed #54. Updated the code generator to support replacing the
service id an version. The code generator currently converts
the ID and the version for the service definition into a
language specific namespace. However, some users may want to
put the id or the version from the JSIDL into the code regardless
of the language. This could allow the service template to be
updated to register the service with a Discovery manager.

These changes add the ability to use the following in the service
and protocol behavior templates:
  * %service_id%
  * %service_version%
  * %service_major_version%
  * %service_minor_version%

See #54.